requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
cleo==0.8.1
click==8.1.6
clikit==0.6.2
crashtest==0.3.1
Faker==19.3.0
fastapi==0.101.0
graphql-core==3.2.3
h11==0.14.0
idna==3.4
inflection==0.5.1
masonite-orm==2.18.6
numpy==1.25.2
pastel==0.2.1
pendulum==2.1.2
psycopg2-binary==2.9.7
pydantic==2.1.1
pydantic_core==2.5.0
pydub==0.25.1
pylev==1.4.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytzdata==2020.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.0
starlette==0.31.0
strawberry-graphql==0.203.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
